Complete all but last section on AboutFunctions
Pushing AboutFunctions.js. All sections with the exception of the last
one are complete. I'm having trouble getting the test to pass for the
"use function body as a string" section.

Trying to figure out the proper output. I think there is a whitespace
issue causing the test to fail.

-JEE