Complete AboutFunctions
All sections are now complete for AboutFunctions.js. I was initially
having trouble with the last section, 'use function body as a string',
however I finally got the correct white-space that matches the correct
output.

NOTE: When converting a function to a string, pay close attention to
the indentation of the function body!

Pushing to my fork.

-JEE